requirements.txt
google-adk
litellm
openai>=1.0
python-dotenv
